Flytt FiltervalgModell til eiga fil slik at det forhåpentlegvis vert lettare å navigere i koden vår etter kvart
