+++ src/filtrering/filtrering-filter/filterform/geografiskbosted-filterform.tsx
@@ -3,7 +3,7 @@ import {isEmptyArray} from 'formik';
 import {MultiSelect} from 'react-multi-select-component';
 import classNames from 'classnames';
 import {Checkbox, CheckboxGroup} from '@navikt/ds-react';
-import {FiltervalgModell} from '../../../model-interfaces';
+import {FiltervalgModell} from '../../../typer/filtervalg-modell';
 import {Grid} from '../../../components/grid/grid';
 import {NullstillKnapp} from '../../../components/nullstill-valg-knapp/nullstill-knapp';
 import {useGeografiskbostedSelector} from '../../../hooks/redux/use-geografiskbosted-selector';
