+++ src/hooks/redux/use-portefolje-selector.tsx
@@ -3,7 +3,8 @@ import {AppState} from '../../reducer';
 import {createSelector} from 'reselect';
 import {getFiltreringState, selectListeVisning} from '../../ducks/ui/listevisning-selectors';
 import {ListevisningState, OversiktType} from '../../ducks/ui/listevisning';
-import {BrukerModell, FiltervalgModell} from '../../model-interfaces';
+import {BrukerModell} from '../../model-interfaces';
+import {FiltervalgModell} from '../../typer/filtervalg-modell';
 import {OrNothing} from '../../utils/types/types';
 import {PortefoljeState} from '../../ducks/portefolje';
 import {EnhettiltakState} from '../../ducks/enhettiltak';
