+++ src/minoversikt/minoversikt-bruker-panel.tsx
@@ -7,7 +7,8 @@ import classNames from 'classnames';
 import {Checkbox, Tag} from '@navikt/ds-react';
 import {BrukerpanelKnapp} from '../components/tabell/brukerpanel-knapp';
 import {Etiketter} from '../components/tabell/etiketter';
-import {BrukerModell, FiltervalgModell} from '../model-interfaces';
+import {BrukerModell} from '../model-interfaces';
+import {FiltervalgModell} from '../typer/filtervalg-modell';
 import {MinOversiktKolonner} from './minoversikt-kolonner';
 import {Kolonne} from '../ducks/ui/listevisning';
 import {logEvent} from '../utils/frontend-logger';
