+++ src/model-interfaces.ts
@@ -1,64 +1,3 @@
-import {FiltreringAktiviteterValg} from './ducks/filtrering';
-
-/* Sortering */
-
-/* Filtervalg*/
-
-/**
- * * * * * VIKTIG! * * * * * VIKTIG! * * * * * VIKTIG! * * * * * VIKTIG! * * * * * VIKTIG! * * * * *
- * Om FiltervalgModell får endringar må ein også oppdatere Portefoljefilter i veilarbfilter.       *
- * Begge repoa må deployast samstundes, elles knekk ein Mine filter i prod.                        *
- *                                                                                                 *
- * Relevant fil: https://github.com/navikt/veilarbfilter/blob/dev/src/main/java/no/nav/pto/veilarbfilter/domene/PortefoljeFilter.java (2024-11-05)
- * Eksempel-PR frå huskelapp: https://github.com/navikt/veilarbfilter/pull/283                     *
- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
-export interface FiltervalgModell {
-    ferdigfilterListe: string[];
-    nyeBrukereForVeileder?: boolean;
-    inaktiveBrukere?: boolean;
-    venterPaSvarFraNAV?: boolean;
-    venterPaSvarFraBruker?: boolean;
-    arbeidslisteKategori: KategoriModell[];
-    alder?: string[];
-    kjonn?: null | string;
-    landgruppe: string[];
-    foedeland: string[];
-    fodselsdagIMnd?: string[];
-    innsatsgruppe?: string[];
-    formidlingsgruppe?: string[];
-    servicegruppe?: string[];
-    veiledere: string[];
-    ytelse: null | string;
-    aktiviteter?: FiltreringAktiviteterValg;
-    aktiviteterForenklet: string[];
-    tiltakstyper: string[];
-    hovedmal?: string[];
-    navnEllerFnrQuery: string;
-    rettighetsgruppe?: string[];
-    manuellBrukerStatus?: string[];
-    veilederNavnQuery: string;
-    registreringstype: string[];
-    cvJobbprofil: null | string;
-    utdanning: string[];
-    utdanningGodkjent: string[];
-    utdanningBestatt: string[];
-    sisteEndringKategori: string[];
-    ulesteEndringer: null | string;
-    tolkebehov: string[];
-    tolkBehovSpraak: string[];
-    stillingFraNavFilter: string[];
-    geografiskBosted: string[];
-    visGeografiskBosted: string[];
-    avvik14aVedtak: string[];
-    ensligeForsorgere: string[];
-    barnUnder18Aar: string[];
-    barnUnder18AarAlder: string[];
-    fargekategorier: FargekategoriModell[];
-    gjeldendeVedtak14a: string[];
-    innsatsgruppeGjeldendeVedtak14a: InnsatsgruppeGjeldendeVedtak14a[];
-    hovedmalGjeldendeVedtak14a: Hovedmal[];
-}
-
 /* Enhet og veileder */
 
 export interface EnhetModell {
